OAUTH2 with Keycloak

+++ springSecurityEazyBank/7eazyBankKeyCloak/src/main/java/com/saha/amit/repository/AccountsRepository.java
@@ -4,11 +4,9 @@ import com.saha.amit.model.Accounts;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-
-
 @Repository
 public interface AccountsRepository extends CrudRepository<Accounts, Long> {
-	
-	Accounts findByCustomerId(int customerId);
+
+    Accounts findByCustomerId(long customerId);
 
 }
