+++ springSecurityEazyBank/7eazyBankKeyCloak/src/main/java/com/saha/amit/repository/LoanRepository.java
@@ -2,17 +2,16 @@ package com.saha.amit.repository;
 
 import java.util.List;
 
-import com.saha.amit.model.Loans;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Repository;
 
-
+import com.saha.amit.model.Loans;
 
 @Repository
 public interface LoanRepository extends CrudRepository<Loans, Long> {
 
 	// @PreAuthorize("hasRole('USER')")
-	List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+	List<Loans> findByCustomerIdOrderByStartDtDesc(long customerId);
 
 }
